+++ visualize-logic/README.md
@@ -3,12 +3,12 @@
 ### Todo
 
 - [ ] Need to do something about there being no "Final" state
-- [ ] Refactor into multiple files
-- [ ] Show actual Txn information
 - [ ] Show what happens when messages go out-of-order
 - [ ] Somehow use for debugging?
+- [x] Animations are bad
+- [x] Refactor into multiple files
+- [x] Show actual Txn information
+- [x] Show information about node in sync
 - [x] Show timeline with highlights; clickable sections
 - [x] Speed for the animations
-- [x] Animations are bad
-- [x] Show information about node in sync
