+++ visualize-logic/src/HydraVis/UI.hs
@@ -1,14 +1,15 @@
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE UndecidableInstances #-}
-{-# OPTIONS_GHC -Wno-missing-local-signatures #-}
-
 -- | The Miso application driving the visualizer UI.
 --
--- The 'Model' is a precomputed list of 'HistoryStep's plus a cursor. Scrubbing
--- the cursor selects a step; the view renders the 'StateChanged' that produced
--- this step, and the resulting 'NodeState' as pretty JSON. No protocol logic
--- runs in 'updateModel' itself; 'HeadLogic.update' was already applied when
--- the history was built (see "HydraVis.History").
+-- This module wires the pieces together: it assembles the page from the panel
+-- modules under "HydraVis.UI" and builds the Miso 'App'. The domain logic
+-- lives in the submodules:
+--
+-- * "HydraVis.UI.Model" — 'Model', 'Action' and pure state accessors.
+-- * "HydraVis.UI.Update" — 'updateModel' / 'applyAction' and subscriptions.
+-- * "HydraVis.UI.Widgets" — shared view helpers.
+-- * "HydraVis.UI.Flow" — the state-transition diagram.
+-- * "HydraVis.UI.Timeline" — the transport toolbar and event markers.
+-- * "HydraVis.UI.Panels" — peers, snapshot, sync, messages, event/state, authoring.
 module HydraVis.UI (
   Model (..),
   Action (..),
@@ -20,157 +21,39 @@ module HydraVis.UI (
   applyInputPure,
 ) where
 
-import Control.Concurrent (forkIO)
-
 import Hydra.Prelude
 
-import Data.Aeson qualified as Aeson
-import Data.Aeson.KeyMap qualified as KM
-import Data.ByteString.Lazy qualified as BSL
-import Data.Map.Strict qualified as Map
-import Data.Set qualified as Set
-import Data.Text qualified as T
-import Data.Text.Encoding qualified as TE
-import Data.Vector (Vector)
-import Data.Vector qualified as V
-import Hydra.Chain.ChainState (ChainSlot (..), IsChainState)
-import Hydra.Events (EventId)
-import Hydra.HeadLogic qualified as HL
+import Hydra.Chain.ChainState (IsChainState)
 import Hydra.HeadLogic.Input (Input)
-import Hydra.HeadLogic.Outcome (Outcome (..), StateChanged (..))
-import Hydra.HeadLogic.State qualified as HS
-import Hydra.HeadLogic.StateEvent (StateEvent (..))
-import Hydra.Ledger (Ledger)
-import Hydra.Network (Host)
-import Hydra.Network qualified as Net
-import Hydra.Node.Environment (Environment)
-import Hydra.Node.Environment qualified as Env
-import Hydra.Node.State (ChainPointTime (..), NodeState (..), SyncedStatus (..), initialChainTime, syncedStatus)
-import Hydra.Tx.HeadParameters qualified as HP
-import Hydra.Tx.IsTx (UTxOType)
-import Hydra.Tx.Party (Party)
-import Hydra.Tx.Snapshot (Snapshot (..), getSnapshot)
-import Hydra.Tx.Snapshot qualified as Snap
-import HydraVis.History (HistoryStep (..), buildHistory, extendHistory, loadEventsAfter)
-import Miso (App (..), Effect, Sub, View, defaultEvents, noEff)
-import Miso.Html (
-  button_,
-  code_,
-  div_,
-  h1_,
-  h2_,
-  input_,
-  onClick,
-  onInput,
-  pre_,
-  span_,
-  text,
-  textarea_,
+import Hydra.HeadLogic.Outcome (StateChanged)
+import Hydra.HeadLogic.StateEvent (StateEvent)
+import Hydra.Node.State (NodeState)
+import HydraVis.UI.Flow (viewFlow)
+import HydraVis.UI.Model (
+  Action (..),
+  AuthoringCtx (..),
+  Model (..),
+  lastSeenIn,
+  mkModel,
  )
-import Miso.Html.Property (max_, min_, rows_, step_, styleInline_, type_, value_)
-import Miso.String (MisoString, ms)
-import Miso.String qualified as MS
-import Miso.Svg.Attribute qualified as SvgA
-import Miso.Svg.Element qualified as Svg
+import HydraVis.UI.Panels (
+  viewAuthoring,
+  viewEvent,
+  viewFanout,
+  viewMessages,
+  viewPeers,
+  viewSnapshot,
+  viewState,
+  viewStuckBanner,
+  viewSync,
+ )
+import HydraVis.UI.Timeline (viewToolbar)
+import HydraVis.UI.Update (applyAction, applyInputPure, followSub, playSub, updateModel)
+import Miso (App (..), View, defaultEvents)
+import Miso.Html (div_, h1_, text)
+import Miso.Html.Property (styleInline_)
+import Miso.Subscription.Keyboard (arrowsSub)
 import Miso.Types (LogLevel (..))
-
-data Model tx = Model
-  { history :: Vector (HistoryStep tx)
-  , cursor :: Int
-  -- ^ Index into the *visible* history (i.e. respects 'hideTicks').
-  , initial :: NodeState tx
-  -- ^ Folded-from state used when the history is empty (i.e. when the first
-  -- batch of events arrives in follow mode against a fresh DB).
-  , follow :: Bool
-  -- ^ Are we currently polling the event store for new rows?
-  , dbPath :: Maybe FilePath
-  -- ^ Echoed in the UI so the user can see what they pointed at.
-  , inputDraft :: MisoString
-  -- ^ Current contents of the authoring textarea.
-  , inputError :: Maybe MisoString
-  -- ^ Last parse/apply error to surface to the user.
-  , ctx :: Maybe (AuthoringCtx tx)
-  -- ^ When 'Just', the UI exposes a panel for authoring 'Input' values and
-  -- stepping 'HeadLogic.update' manually.
-  , hideTicks :: Bool
-  -- ^ When true, 'TickObserved' events are filtered out of the time
-  -- slider, the JSON panes, and the state-transition counts. Real Hydra
-  -- logs are otherwise dominated by ticks.
-  , playing :: Bool
-  -- ^ Auto-advance the cursor once per 'playSub' tick.
-  , speedIdx :: Int
-  -- ^ Index into 'speedTable' controlling playback speed.
-  , playFrame :: Int
-  -- ^ Counts 'playSub' ticks since the last advance; used to slow playback
-  -- below the base tick rate (see 'speedTable').
-  }
-  deriving stock (Generic)
-
-deriving stock instance (IsChainState tx, Eq (NodeState tx)) => Eq (Model tx)
-
--- | Everything 'HeadLogic.update' needs to run a step that does not come out
--- of a SQLite file. Two of the three fields are first-class values
--- ('Environment' and 'NodeState'); 'Ledger' is a record-of-functions so we
--- treat it as opaque for diffing.
-data AuthoringCtx tx = AuthoringCtx
-  { ctxEnvironment :: Environment
-  , ctxLedger :: Ledger tx
-  , ctxPresets :: [(MisoString, Input tx)]
-  -- ^ Named pre-canned inputs the UI shows as one-click buttons.
-  }
-
--- | Always-equal so that the Miso diff does not bail on us; the ctx is set
--- once at startup and never changes.
-instance Eq (AuthoringCtx tx) where
-  _ == _ = True
-
-data Action tx
-  = NoOp
-  | Next
-  | Prev
-  | GoFirst
-  | GoLast
-  | SetCursor MisoString
-  | AppendEvents [StateEvent tx]
-  | SetInputDraft MisoString
-  | LoadPreset MisoString
-  | ApplyDraft
-  | ToggleHideTicks
-  | TogglePlay
-  | SetSpeed Int
-  | -- | Fired by 'playSub' once per tick; the update handler decides whether
-    -- to advance based on 'playing'.
-    MaybeAdvance
-  deriving stock (Generic)
-
-mkModel ::
-  Maybe (AuthoringCtx tx) ->
-  Maybe FilePath ->
-  Bool ->
-  NodeState tx ->
-  Int ->
-  [HistoryStep tx] ->
-  Model tx
-mkModel authoring path isFollowing initialState startCursor steps =
-  let hs = V.fromList steps
-      hideT = True
-      visN =
-        V.length $
-          if hideT then V.filter (not . isTickEvent . event) hs else hs
-   in Model
-        { history = hs
-        , cursor = max 0 (min (max 0 (visN - 1)) startCursor)
-        , initial = initialState
-        , follow = isFollowing
-        , dbPath = path
-        , inputDraft = ""
-        , inputError = Nothing
-        , ctx = authoring
-        , hideTicks = hideT
-        , playing = False
-        , speedIdx = defaultSpeedIdx
-        , playFrame = 0
-        }
 
 mkApp ::
   ( IsChainState tx
@@ -190,6 +73,7 @@ mkApp initialModel =
     , view = viewModel
     , subs =
         playSub
+          : arrowsSub ArrowSeek
           : case (follow initialModel, dbPath initialModel) of
             (True, Just path) -> [followSub path (lastSeenIn initialModel)]
             _ -> []
@@ -197,259 +81,6 @@ mkApp initialModel =
     , mountPoint = Nothing
     , logLevel = Off
     }
-
-lastSeenIn :: Model tx -> Maybe EventId
-lastSeenIn m =
-  if V.null (history m)
-    then Nothing
-    else Just (eventId (event (V.last (history m))))
-
-updateModel ::
-  forall tx.
-  (IsChainState tx, FromJSON (Input tx)) =>
-  Action tx ->
-  Model tx ->
-  Effect (Action tx) (Model tx)
-updateModel a m = case a of
-  NoOp -> noEff m
-  Next -> noEff $ m{cursor = clamp m (cursor m + 1)}
-  Prev -> noEff $ m{cursor = clamp m (cursor m - 1)}
-  GoFirst -> noEff $ m{cursor = 0}
-  GoLast -> noEff $ m{cursor = lastIndex m}
-  SetCursor s ->
-    case readMaybe (MS.unpack s) :: Maybe Int of
-      Just n -> noEff $ m{cursor = clamp m n}
-      Nothing -> noEff m
-  AppendEvents events ->
-    let existing = V.toList (history m)
-        extended = extendHistory (initial m) existing events
-        followingTail = cursor m == max 0 (V.length (visibleHistory m) - 1)
-        newModel = m{history = V.fromList extended}
-        newCursor =
-          if followingTail
-            then max 0 (V.length (visibleHistory newModel) - 1)
-            else cursor m
-     in noEff newModel{cursor = newCursor}
-  SetInputDraft s -> noEff $ m{inputDraft = s, inputError = Nothing}
-  LoadPreset s -> noEff $ m{inputDraft = s, inputError = Nothing}
-  ApplyDraft -> applyDraft m
-  ToggleHideTicks -> noEff $ m{hideTicks = not (hideTicks m), cursor = 0}
-  TogglePlay -> noEff $ m{playing = not (playing m), playFrame = 0}
-  SetSpeed i -> noEff $ m{speedIdx = clampSpeed i, playFrame = 0}
-  MaybeAdvance -> noEff (stepPlayback (lastIndex m) m)
- where
-  clamp :: Model tx -> Int -> Int
-  clamp model n = max 0 (min (lastIndex model) n)
-  lastIndex :: Model tx -> Int
-  lastIndex model = max 0 (V.length (visibleHistory model) - 1)
-
--- | The history vector after the optional Tick filter.
-visibleHistory :: Model tx -> Vector (HistoryStep tx)
-visibleHistory m
-  | hideTicks m = V.filter (not . isTickEvent . event) (history m)
-  | otherwise = history m
-
-isTickEvent :: StateEvent tx -> Bool
-isTickEvent e = case stateChanged e of
-  TickObserved{} -> True
-  _ -> False
-
--- | A periodic tick that drives the Play button. Always installed; the
--- update handler checks 'playing' and decides whether to advance. Ticks at the
--- base rate; 'speedTable' slows playback below this by skipping ticks.
-playSub :: Sub (Action tx)
-playSub sink =
-  liftIO $
-    void $
-      forkIO $
-        forever $ do
-          threadDelay 0.2
-          sink MaybeAdvance
-
--- | Playback speeds: @(label, ticksPerStep, stride)@. With the 0.2s base tick,
--- @ticksPerStep@ stretches the interval (for slow motion) and @stride@ jumps
--- multiple events per step (for fast-forward).
-speedTable :: [(MisoString, Int, Int)]
-speedTable =
-  [ ("0.5x", 4, 1)
-  , ("1x", 2, 1)
-  , ("2x", 1, 1)
-  , ("4x", 1, 2)
-  , ("8x", 1, 4)
-  ]
-
-defaultSpeedIdx :: Int
-defaultSpeedIdx = 1
-
-clampSpeed :: Int -> Int
-clampSpeed = max 0 . min (length speedTable - 1)
-
-speedParams :: Int -> (Int, Int)
-speedParams i = case drop (clampSpeed i) speedTable of
-  ((_, period, stride) : _) -> (period, stride)
-  [] -> (1, 1)
-
--- | One 'playSub' tick of playback: advance the cursor when enough ticks have
--- elapsed for the current speed, stopping (and clearing 'playing') at the end.
-stepPlayback :: Int -> Model tx -> Model tx
-stepPlayback lastIndex m
-  | not (playing m) = m
-  | otherwise =
-      let (period, stride) = speedParams (speedIdx m)
-          frame = playFrame m + 1
-       in if frame < period
-            then m{playFrame = frame}
-            else
-              let next = cursor m + stride
-               in if next >= lastIndex
-                    then m{cursor = lastIndex, playing = False, playFrame = 0}
-                    else m{cursor = next, playFrame = 0}
-
-applyDraft ::
-  (IsChainState tx, FromJSON (Input tx)) =>
-  Model tx ->
-  Effect (Action tx) (Model tx)
-applyDraft m = case ctx m of
-  Nothing ->
-    noEff m{inputError = Just "no authoring context (run with --simple)"}
-  Just AuthoringCtx{ctxEnvironment, ctxLedger} ->
-    case Aeson.eitherDecodeStrict' (encodeUtf8 (MS.fromMisoString (inputDraft m) :: Text)) of
-      Left err ->
-        noEff m{inputError = Just (MS.ms ("parse error: " <> err))}
-      Right input -> noEff (applyInput ctxEnvironment ctxLedger input m)
-
--- | Pure version of the @ApplyDraft@ branch of 'updateModel': run
--- 'HeadLogic.update' on @input@ against the state at the cursor (or the
--- model's initial state when the history is empty) and append the resulting
--- 'StateChanged's as new 'HistoryStep's.
-applyInputPure ::
-  IsChainState tx =>
-  Environment ->
-  Ledger tx ->
-  Input tx ->
-  Model tx ->
-  Model tx
-applyInputPure = applyInput
-
-applyInput ::
-  forall tx.
-  IsChainState tx =>
-  Environment ->
-  Ledger tx ->
-  Input tx ->
-  Model tx ->
-  Model tx
-applyInput env ledger input m =
-  let priorState = case nonEmpty (V.toList (history m)) of
-        -- We append to the *raw* history, so the prior state is the last raw
-        -- entry (not the visible one), which respects all events including
-        -- the ticks the user may have filtered out.
-        Just ne -> stateAfter (last ne)
-        Nothing -> initial m
-      nextEventId :: Word64
-      nextEventId = case nonEmpty (V.toList (history m)) of
-        Nothing -> 0
-        Just ne -> 1 + eventId (event (last ne))
-      walltime =
-        addUTCTime
-          (fromIntegral (V.length (history m)))
-          initialChainTime
-      outcome :: Outcome tx
-      outcome = HL.update env ledger walltime priorState input
-      synthEvents = case outcome of
-        Continue{stateChanges} -> withIds nextEventId walltime stateChanges
-        Wait{stateChanges} -> withIds nextEventId walltime stateChanges
-        Error{} -> []
-      newSteps = buildHistory priorState synthEvents
-      err = case outcome of
-        Error{error = e} ->
-          Just (MS.ms ("update returned Error: " <> show (e :: HL.LogicError tx) :: Text))
-        _ -> Nothing
-   in m
-        { history = history m <> V.fromList newSteps
-        , cursor =
-            if null newSteps
-              then cursor m
-              else V.length (history m) + length newSteps - 1
-        , inputError = err
-        , inputDraft = if isNothing err then "" else inputDraft m
-        }
- where
-  withIds startId t =
-    zipWith
-      (\i sc -> StateEvent{eventId = i, stateChanged = sc, time = t})
-      [startId ..]
-
--- | Pure variant of the action handler used by multi-party views that don't
--- want to thread 'Miso.Effect'. ApplyDraft falls through to 'applyInput' when
--- the textarea parses; everything else is straightforward field tweaks.
-applyAction ::
-  forall tx.
-  (IsChainState tx, FromJSON (Input tx)) =>
-  Action tx ->
-  Model tx ->
-  Model tx
-applyAction a m = case a of
-  NoOp -> m
-  Next -> m{cursor = clamp (cursor m + 1)}
-  Prev -> m{cursor = clamp (cursor m - 1)}
-  GoFirst -> m{cursor = 0}
-  GoLast -> m{cursor = lastIndex}
-  SetCursor s -> case readMaybe (MS.unpack s) :: Maybe Int of
-    Just n -> m{cursor = clamp n}
-    Nothing -> m
-  AppendEvents events ->
-    let existing = V.toList (history m)
-        extended = extendHistory (initial m) existing events
-        followingTail = cursor m == max 0 (V.length (visibleHistory m) - 1)
-        newModel = m{history = V.fromList extended}
-        newCursor =
-          if followingTail
-            then max 0 (V.length (visibleHistory newModel) - 1)
-            else cursor m
-     in newModel{cursor = newCursor}
-  SetInputDraft s -> m{inputDraft = s, inputError = Nothing}
-  LoadPreset s -> m{inputDraft = s, inputError = Nothing}
-  ApplyDraft -> case ctx m of
-    Nothing -> m{inputError = Just "no authoring context"}
-    Just AuthoringCtx{ctxEnvironment, ctxLedger} ->
-      case Aeson.eitherDecodeStrict' (encodeUtf8 (MS.fromMisoString (inputDraft m) :: Text)) of
-        Left err -> m{inputError = Just (MS.ms ("parse error: " <> err))}
-        Right input -> applyInput ctxEnvironment ctxLedger input m
-  ToggleHideTicks -> m{hideTicks = not (hideTicks m), cursor = 0}
-  TogglePlay -> m{playing = not (playing m), playFrame = 0}
-  SetSpeed i -> m{speedIdx = clampSpeed i, playFrame = 0}
-  MaybeAdvance -> stepPlayback lastIndex m
- where
-  lastIndex = max 0 (V.length (visibleHistory m) - 1)
-  clamp n = max 0 (min lastIndex n)
-
--- | Background poll loop. Reads any new rows from the SQLite file at a
--- fixed cadence and dispatches them as 'AppendEvents'.
-followSub ::
-  FromJSON (StateEvent tx) =>
-  FilePath ->
-  Maybe EventId ->
-  Sub (Action tx)
-followSub path initialLastSeen sink =
-  liftIO $ do
-    seenRef <- newIORef initialLastSeen
-    _ <- forkIO $ forever $ do
-      threadDelay 1.5
-      seen <- readIORef seenRef
-      newEvents <- loadEventsAfter path seen
-      case nonEmpty newEvents of
-        Nothing -> pure ()
-        Just ne -> do
-          modifyIORef' seenRef (\_ -> Just (eventId (last ne)))
-          putTextLn $
-            "follow: picked up "
-              <> show (length ne)
-              <> " new events (last eventId = "
-              <> show (eventId (last ne))
-              <> ")"
-          sink (AppendEvents (toList ne))
-    pure ()
 
 viewModel ::
   (IsChainState tx, ToJSON (Input tx), ToJSON (NodeState tx), ToJSON (StateChanged tx)) =>
@@ -460,1317 +91,14 @@ viewModel m =
     [styleInline_ "font-family: system-ui, sans-serif; max-width: 900px; margin: 1rem auto; padding: 0 1rem;"]
     [ h1_ [] [text "hydra-vis"]
     , viewToolbar m
+    , viewStuckBanner m
     , viewFlow m
     , viewPeers m
     , viewSync m
     , viewMessages m
     , viewSnapshot m
+    , viewFanout m
     , viewAuthoring m
     , viewEvent m
     , viewState m
     ]
-
-viewToolbar :: Model tx -> View (Action tx)
-viewToolbar m =
-  div_
-    [styleInline_ "margin-bottom: 1rem;"]
-    [ div_
-        [styleInline_ "display: flex; gap: 0.5rem; align-items: center; flex-wrap: wrap; margin-bottom: 0.4rem;"]
-        ( [ button_ [onClick GoFirst] [text "|<"]
-          , button_ [onClick Prev] [text "<"]
-          , button_
-              [ onClick TogglePlay
-              , styleInline_ $
-                  "padding: 0.25rem 0.6rem; "
-                    <> if playing m
-                      then "background: #ffd8d8; border: 1px solid #c33;"
-                      else "background: #d8ffd8; border: 1px solid #393;"
-              ]
-              [text (if playing m then "Pause" else "Play")]
-          , button_ [onClick Next] [text ">"]
-          , button_ [onClick GoLast] [text ">|"]
-          , span_ [styleInline_ "color: #888; margin-left: 0.5rem;"] [text "speed"]
-          ]
-            <> speedButtons
-            <> [ button_
-                  [ onClick ToggleHideTicks
-                  , styleInline_ $
-                      "margin-left: 0.5rem; padding: 0.25rem 0.6rem; "
-                        <> if hideTicks m
-                          then "background: #cce5ff; border: 1px solid #6ab;"
-                          else "background: #fff; border: 1px solid #ccc;"
-                  ]
-                  [text (if hideTicks m then "Show Ticks" else "Hide Ticks")]
-               , span_
-                  [styleInline_ "margin-left: auto; min-width: 7ch; text-align: right; font-variant-numeric: tabular-nums;"]
-                  [text $ ms (show oneBased :: Text) <> "/" <> ms (show totalSteps :: Text)]
-               ]
-        )
-    , viewTimeline m
-    , input_
-        [ type_ "range"
-        , min_ "0"
-        , max_ (ms (show (max 0 (totalSteps - 1)) :: Text))
-        , step_ "1"
-        , value_ (ms (show (cursor m) :: Text))
-        , onInput SetCursor
-        , styleInline_ "width: 100%; box-sizing: border-box; margin: 0;"
-        ]
-    , markerLegend
-    ]
- where
-  totalSteps :: Int
-  totalSteps = V.length (visibleHistory m)
-  oneBased :: Int
-  oneBased = if totalSteps == 0 then 0 else cursor m + 1
-  speedButtons :: [View (Action tx)]
-  speedButtons =
-    [ button_
-      [ onClick (SetSpeed i)
-      , styleInline_ $
-          "padding: 0.15rem 0.4rem; font-size: 0.8rem; "
-            <> if i == speedIdx m
-              then "background: #cce5ff; border: 1px solid #6ab;"
-              else "background: #fff; border: 1px solid #ccc;"
-      ]
-      [text lbl]
-    | (i, (lbl, _, _)) <- zip [0 ..] speedTable
-    ]
-
--- | Marker kinds surfaced as labelled lines on the timeline.
-data MarkerKind = MOpen | MIncrement | MDecrement | MClose | MFanout | MPeerUp | MPeerDown
-  deriving stock (Eq, Show)
-
--- | Classify an event as a "big" timeline marker, if it is one.
-markerOf :: StateChanged tx -> Maybe MarkerKind
-markerOf = \case
-  HeadOpened{} -> Just MOpen
-  CommitFinalized{} -> Just MIncrement
-  DecommitFinalized{} -> Just MDecrement
-  HeadClosed{} -> Just MClose
-  HeadFannedOut{} -> Just MFanout
-  PeerConnected{} -> Just MPeerUp
-  PeerDisconnected{} -> Just MPeerDown
-  _ -> Nothing
-
-markerColor :: MarkerKind -> MisoString
-markerColor = \case
-  MOpen -> "#06c"
-  MIncrement -> "#393"
-  MDecrement -> "#c66"
-  MClose -> "#a33"
-  MFanout -> "#c83"
-  MPeerUp -> "#27a"
-  MPeerDown -> "#999"
-
-markerLabel :: MarkerKind -> MisoString
-markerLabel = \case
-  MOpen -> "open"
-  MIncrement -> "increment"
-  MDecrement -> "decrement"
-  MClose -> "close"
-  MFanout -> "fanout"
-  MPeerUp -> "peer joined"
-  MPeerDown -> "peer left"
-
--- | The big events along the (Tick-filtered) timeline, by slider index.
-bigEvents :: Vector (HistoryStep tx) -> [(Int, MarkerKind)]
-bigEvents steps =
-  [(i, k) | (i, s) <- zip [0 ..] (V.toList steps), Just k <- [markerOf (stateChanged (event s))]]
-
--- | A horizontal track above the slider with a coloured line at each big event
--- and a blue marker at the current cursor, so notable moments are visible at a
--- glance and line up with the slider below.
-viewTimeline :: Model tx -> View (Action tx)
-viewTimeline m =
-  let n = V.length (visibleHistory m)
-      xOf i = if n <= 1 then 0 else (i * 1000) `div` (n - 1)
-      markers = bigEvents (visibleHistory m)
-      cur = xOf (cursor m)
-   in Svg.svg_
-        [ SvgA.viewBox_ "0 0 1000 36"
-        , SvgA.width_ "100%"
-        , SvgA.preserveAspectRatio_ "none"
-        , styleInline_ "display: block; height: 34px;"
-        ]
-        ( Svg.rect_
-            [ SvgA.x_ "0"
-            , SvgA.y_ "14"
-            , SvgA.width_ "1000"
-            , SvgA.height_ "8"
-            , SvgA.fill_ "#eee"
-            ]
-            []
-            : [ Svg.line_
-                [ SvgA.x1_ (mInt (xOf i))
-                , SvgA.y1_ "2"
-                , SvgA.x2_ (mInt (xOf i))
-                , SvgA.y2_ "34"
-                , SvgA.stroke_ (markerColor k)
-                , SvgA.strokeWidth_ "2"
-                ]
-                []
-              | (i, k) <- markers
-              ]
-              <> [ Svg.line_
-                    [ SvgA.x1_ (mInt cur)
-                    , SvgA.y1_ "0"
-                    , SvgA.x2_ (mInt cur)
-                    , SvgA.y2_ "36"
-                    , SvgA.stroke_ "#06c"
-                    , SvgA.strokeWidth_ "3"
-                    ]
-                    []
-                 ]
-        )
-
--- | Colour key for the timeline markers.
-markerLegend :: View (Action tx)
-markerLegend =
-  div_
-    [styleInline_ "display: flex; gap: 0.9rem; flex-wrap: wrap; margin-top: 0.35rem; font-size: 0.78rem; color: #555;"]
-    [ span_
-      [styleInline_ "display: inline-flex; align-items: center; gap: 0.3rem;"]
-      [ span_
-          [styleInline_ ("display: inline-block; width: 10px; height: 10px; border-radius: 2px; background: " <> markerColor k <> ";")]
-          []
-      , text (markerLabel k)
-      ]
-    | k <- [MOpen, MIncrement, MDecrement, MClose, MFanout, MPeerUp, MPeerDown]
-    ]
-
-viewEvent :: ToJSON (StateChanged tx) => Model tx -> View (Action tx)
-viewEvent m = case currentStep m of
-  Nothing ->
-    div_
-      [styleInline_ panelStyle]
-      [ h2_ [] [text "no events loaded"]
-      , text "Pass --db PATH or --gen-db PATH on startup."
-      ]
-  Just step ->
-    div_
-      [styleInline_ panelStyle]
-      [ h2_ [] [text "event"]
-      , span_
-          [styleInline_ "color: #555;"]
-          [ text $
-              "eventId="
-                <> ms (show (eventId (event step)) :: Text)
-                <> " time="
-                <> ms (show (time (event step)) :: Text)
-          ]
-      , renderJson (stateChanged (event step))
-      ]
-
--- | Surface the head's current snapshot so it's immediately visible without
--- having to scroll through the full 'NodeState' JSON. Updates live as
--- 'PartySignedSnapshot' events accumulate, because each one mutates the
--- 'seenSnapshot' record we read from here.
-viewSnapshot :: IsChainState tx => Model tx -> View (Action tx)
-viewSnapshot m = case currentStep m of
-  Nothing -> div_ [] []
-  Just step ->
-    div_
-      [styleInline_ panelStyle]
-      [ h2_ [] [text "current snapshot"]
-      , snapshotBody (headState (stateAfter step))
-      ]
-
-snapshotBody :: forall tx. IsChainState tx => HS.HeadState tx -> View (Action tx)
-snapshotBody = \case
-  HS.Idle _ ->
-    italic "head is Idle; no snapshot yet."
-  HS.Open HS.OpenState{HS.coordinatedHeadState = chs} ->
-    let HS.CoordinatedHeadState{HS.seenSnapshot = seen, HS.confirmedSnapshot = cs} = chs
-        snap = getSnapshot @tx cs
-     in div_
-          []
-          [ row "confirmed snapshot #" (showT (number snap))
-          , headBalance snap
-          , seenSnapshotRows seen
-          ]
-  HS.Closed HS.ClosedState{HS.confirmedSnapshot = cs} ->
-    let snap = getSnapshot @tx cs
-     in div_
-          []
-          [ row "confirmed snapshot #" (showT (number snap))
-          , row "confirmed txs" (showT (length (confirmed snap)))
-          , headBalance snap
-          ]
- where
-  showT :: Show a => a -> Text
-  showT = show
-  -- The lovelace held in the head and to whom it is allocated, plus any
-  -- pending deposit/decommit, derived from the confirmed snapshot's UTxO.
-  headBalance :: Snapshot tx -> View (Action tx)
-  headBalance snap =
-    utxoSummaryView @tx (Snap.utxo snap) (Snap.utxoToCommit snap) (Snap.utxoToDecommit snap)
-
-seenSnapshotRows :: HS.SeenSnapshot tx -> View (Action tx)
-seenSnapshotRows = \case
-  HS.NoSeenSnapshot ->
-    row "in flight" "none"
-  HS.LastSeenSnapshot{HS.lastSeen} ->
-    div_
-      []
-      [row "last seen #" (show lastSeen :: Text), row "in flight" "none"]
-  HS.RequestedSnapshot{HS.lastSeen, HS.requested} ->
-    div_
-      []
-      [ row "last seen #" (show lastSeen :: Text)
-      , row "requested #" (show requested :: Text)
-      , row "in flight" "ReqSn sent, awaiting acks"
-      ]
-  HS.SeenSnapshot{HS.snapshot = snap, HS.signatories} ->
-    div_
-      []
-      [ row "in flight #" (show (number snap) :: Text)
-      , row "signatories" (sigSummary signatories)
-      , row "transactions" (show (length (confirmed snap)) :: Text)
-      ]
- where
-  sigSummary :: Map.Map Party a -> Text
-  sigSummary m =
-    let n = Map.size m
-        names = T.intercalate ", " (shortParty <$> Map.keys m)
-     in show n <> "  [" <> names <> "]"
-
-row :: MisoString -> Text -> View (Action tx)
-row k v =
-  div_
-    [styleInline_ "display: flex; gap: 0.75rem; margin: 0.15rem 0;"]
-    [ span_ [styleInline_ "min-width: 12rem; color: #666;"] [text k]
-    , span_ [styleInline_ "font-family: ui-monospace, monospace;"] [text (ms v)]
-    ]
-
-italic :: MisoString -> View (Action tx)
-italic t = div_ [styleInline_ "color: #777; font-style: italic;"] [text t]
-
--- * Peers
-
--- | Surface who is in the head: the parties from the head parameters (plus
--- the head id and contestation period), marking our own party when we know it
--- from an 'Environment'. Before the head opens there are no parties on-chain,
--- so we fall back to whatever the 'Environment' configured.
-viewPeers :: Model tx -> View (Action tx)
-viewPeers m = case currentStep m of
-  Nothing -> div_ [] []
-  Just step ->
-    div_
-      [styleInline_ panelStyle]
-      [ h2_ [] [text "peers"]
-      , peersBody (ctx m) (headState (stateAfter step))
-      ]
-
-peersBody :: Maybe (AuthoringCtx tx) -> HS.HeadState tx -> View (Action tx)
-peersBody mctx = \case
-  HS.Idle _ -> case mctx of
-    Nothing -> italic "participants are known once the head opens."
-    Just c -> fromEnv (ctxEnvironment c)
-  HS.Open HS.OpenState{HS.parameters = ps, HS.headId = hid} -> openClosed ps hid
-  HS.Closed HS.ClosedState{HS.parameters = ps, HS.headId = hid} -> openClosed ps hid
- where
-  me = Env.party . ctxEnvironment <$> mctx
-
-  openClosed ps hid =
-    div_
-      []
-      [ row "head id" (shortText (jsonStringy hid))
-      , row "contestation period" (show (HP.contestationPeriod ps))
-      , row "parties" (show (length (HP.parties ps)))
-      , partyList me (HP.parties ps)
-      ]
-
-  fromEnv env =
-    div_
-      []
-      [ row "configured peers" (Env.configuredPeers env)
-      , row "parties" (show (1 + length (Env.otherParties env)))
-      , partyList (Just (Env.party env)) (Env.party env : Env.otherParties env)
-      ]
-
-partyList :: Maybe Party -> [Party] -> View (Action tx)
-partyList me parties =
-  div_
-    [styleInline_ "margin-top: 0.25rem;"]
-    (map partyRow parties)
- where
-  partyRow p =
-    div_
-      [styleInline_ "display: flex; gap: 0.5rem; margin: 0.1rem 0 0.1rem 1rem; font-family: ui-monospace, monospace;"]
-      ( span_ [] [text (ms (shortParty p))]
-          : [span_ [styleInline_ "color: #06c;"] [text "(you)"] | Just p == me]
-      )
-
-shortParty :: Party -> Text
-shortParty p = T.take 16 (T.drop (T.length ("Party {vkey = \"" :: Text)) (show p))
-
--- * Head balance
-
--- | A breakdown of a snapshot UTxO. Derived by walking the UTxO's JSON so it
--- works for any tx type without a concrete-type constraint: Cardano UTxOs
--- serialize to an object keyed by input with per-output @address@ and
--- @value.lovelace@; value-less ledgers (e.g. SimpleTx) serialize to an array,
--- in which case 'usValued' is False and we only know the entry count.
-data UtxoSummary = UtxoSummary
-  { usEntries :: Int
-  , usValued :: Bool
-  , usTotalLovelace :: Integer
-  , usByAddress :: [(Text, Integer)]
-  }
-
-summariseUtxo :: forall tx. ToJSON (UTxOType tx) => UTxOType tx -> UtxoSummary
-summariseUtxo u = case Aeson.toJSON u of
-  Aeson.Object o ->
-    let perAddr = [pair | v <- KM.elems o, Just pair <- [entryAddrLovelace v]]
-        byAddr = Map.toList (Map.fromListWith (+) perAddr)
-     in UtxoSummary
-          { usEntries = KM.size o
-          , usValued = True
-          , usTotalLovelace = sum (snd <$> perAddr)
-          , usByAddress = sortOn (negate . snd) byAddr
-          }
-  Aeson.Array a ->
-    UtxoSummary{usEntries = length a, usValued = False, usTotalLovelace = 0, usByAddress = []}
-  _ -> UtxoSummary{usEntries = 0, usValued = False, usTotalLovelace = 0, usByAddress = []}
- where
-  entryAddrLovelace = \case
-    Aeson.Object eo -> do
-      addr <- KM.lookup "address" eo >>= asText
-      let love = fromMaybe 0 (KM.lookup "value" eo >>= asObject >>= KM.lookup "lovelace" >>= asInteger)
-      pure (addr, love)
-    _ -> Nothing
-  asText = \case Aeson.String t -> Just t; _ -> Nothing
-  asObject = \case Aeson.Object o -> Just o; _ -> Nothing
-  asInteger = \case Aeson.Number n -> Just (round n); _ -> Nothing
-
-utxoSummaryView ::
-  forall tx.
-  ToJSON (UTxOType tx) =>
-  UTxOType tx ->
-  Maybe (UTxOType tx) ->
-  Maybe (UTxOType tx) ->
-  View (Action tx)
-utxoSummaryView u toCommit toDecommit =
-  let s = summariseUtxo @tx u
-   in div_
-        [styleInline_ "margin-top: 0.4rem;"]
-        ( [ row "outputs in head" (show (usEntries s))
-          ]
-            <> ( if usValued s
-                  then
-                    [ row "total in head" (lovelaceText (usTotalLovelace s))
-                    , allocationBlock (usByAddress s)
-                    ]
-                  else [italic "value-less ledger (SimpleTx); no lovelace to show."]
-               )
-            <> pendingRow "pending deposit (in)" toCommit
-            <> pendingRow "pending decommit (out)" toDecommit
-        )
- where
-  pendingRow lbl = \case
-    Nothing -> []
-    Just pu ->
-      let ps = summariseUtxo @tx pu
-          detail =
-            show (usEntries ps)
-              <> " outputs"
-              <> (if usValued ps then ", " <> lovelaceText (usTotalLovelace ps) else "")
-       in [row lbl detail]
-
-allocationBlock :: [(Text, Integer)] -> View (Action tx)
-allocationBlock [] = div_ [] []
-allocationBlock xs =
-  div_
-    [styleInline_ "margin-top: 0.25rem;"]
-    ( span_ [styleInline_ "color: #666;"] [text "allocation"]
-        : map addrRow xs
-    )
- where
-  addrRow (addr, love) =
-    div_
-      [styleInline_ "display: flex; gap: 0.75rem; margin: 0.1rem 0 0.1rem 1rem;"]
-      [ span_
-          [styleInline_ "min-width: 22rem; color: #444; font-family: ui-monospace, monospace;"]
-          [text (ms (shortAddr addr))]
-      , span_
-          [styleInline_ "font-family: ui-monospace, monospace;"]
-          [text (ms (lovelaceText love))]
-      ]
-
--- | Render lovelace with thousands separators alongside the ADA value.
-lovelaceText :: Integer -> Text
-lovelaceText n = commas n <> " lovelace (" <> adaText n <> " ADA)"
-
-adaText :: Integer -> Text
-adaText n =
-  let (q, r) = abs n `divMod` 1_000_000
-      frac = T.dropWhileEnd (== '0') (T.justifyRight 6 '0' (show r))
-      frac' = if T.null frac then "0" else frac
-   in (if n < 0 then "-" else "") <> commas q <> "." <> frac'
-
-commas :: Integer -> Text
-commas n =
-  let digits = show (abs n) :: Text
-      chunks = map T.reverse (chunk3 (T.reverse digits))
-   in (if n < 0 then "-" else "") <> T.intercalate "," (reverse chunks)
- where
-  chunk3 t
-    | T.null t = []
-    | otherwise = T.take 3 t : chunk3 (T.drop 3 t)
-
-shortAddr :: Text -> Text
-shortAddr a
-  | T.length a <= 24 = a
-  | otherwise = T.take 14 a <> "..." <> T.takeEnd 6 a
-
-shortText :: Text -> Text
-shortText t
-  | T.length t <= 28 = t
-  | otherwise = T.take 24 t <> "..."
-
--- | Render a value whose JSON is a plain string (e.g. 'HeadId', which @show@s
--- as raw escaped bytes) via that string. Falls back to compact JSON otherwise.
-jsonStringy :: ToJSON a => a -> Text
-jsonStringy x = case Aeson.toJSON x of
-  Aeson.String t -> t
-  other -> TE.decodeUtf8 (BSL.toStrict (Aeson.encode other))
-
--- * Node sync
-
--- | Show whether the node is caught up with the chain, the slot/time it last
--- observed, and how far its clock has drifted, plus running counts of the
--- chain/connectivity events seen up to the cursor.
-viewSync :: Model tx -> View (Action tx)
-viewSync m = case currentStep m of
-  Nothing -> div_ [] []
-  Just step ->
-    let counts = countActivity (historyUpToCursor m)
-        ns = stateAfter step
-        ChainPointTime{currentSlot = ChainSlot slot, currentChainTime, drift} =
-          chainPointTime ns
-        (statusText, statusColor) = case syncedStatus ns of
-          InSync -> ("in sync", "#2a7")
-          CatchingUp -> ("catching up", "#c83")
-     in div_
-          [styleInline_ panelStyle]
-          [ h2_ [] [text "node sync"]
-          , div_
-              [styleInline_ "display: flex; gap: 0.75rem; margin: 0.15rem 0;"]
-              [ span_ [styleInline_ "min-width: 12rem; color: #666;"] [text "status"]
-              , span_
-                  [styleInline_ ("font-weight: 600; color: " <> statusColor <> ";")]
-                  [text statusText]
-              ]
-          , row "current slot" (show slot)
-          , row "chain time" (show currentChainTime)
-          , row "clock drift" (show drift)
-          , div_
-              [styleInline_ "display: flex; gap: 0.75rem; margin: 0.15rem 0;"]
-              [ span_ [styleInline_ "min-width: 12rem; color: #666;"] [text "network"]
-              , let (t, col) = case acConnected counts of
-                      Just True -> ("connected", "#2a7")
-                      Just False -> ("not connected", "#a33")
-                      Nothing -> ("unknown", "#999")
-                 in span_ [styleInline_ ("font-weight: 600; color: " <> col <> ";")] [text t]
-              ]
-          , row "peers connected" (connectedPeersText (acConnectedPeers counts))
-          , row "ticks observed" (show (acTicks counts))
-          , row "rollbacks" (show (acRollbacks counts))
-          , row "sync transitions" (show (acSynced counts) <> " synced, " <> show (acUnsynced counts) <> " unsynced")
-          ]
-
--- * Network messages
-
--- | Network message activity, grouped by peer where the log records a sender.
--- Raw network messages are not persisted; we count the 'StateChanged' each
--- message produces when applied. Only @AckSn@ (-> 'PartySignedSnapshot')
--- carries its sender, so that is the one we can attribute per peer; the rest
--- (@ReqTx@, @ReqSn@, @ReqDec@) are shown as node-wide totals with a note.
-viewMessages :: Model tx -> View (Action tx)
-viewMessages m = case currentStep m of
-  Nothing -> div_ [] []
-  Just step ->
-    let c = countActivity (historyUpToCursor m)
-        parties = headParties (headState (stateAfter step))
-        peers = if null parties then Map.keys (acAckByParty c) else parties
-     in div_
-          [styleInline_ panelStyle]
-          [ h2_ [] [text "network messages"]
-          , div_
-              [styleInline_ "color: #888; font-size: 0.85rem; margin-bottom: 0.5rem;"]
-              [ text
-                  "Counted from the persisted state changes each message \
-                  \produces; raw sends are not logged. Only AckSn records its \
-                  \sender, so it is grouped per peer below; the rest are \
-                  \node-wide received counts."
-              ]
-          , span_ [styleInline_ "color: #666;"] [text "AckSn (snapshot signatures) by peer"]
-          , peerAcks peers (acAckByParty c)
-          , div_ [styleInline_ "margin-top: 0.6rem; color: #666;"] [text "node-wide (sender not recorded)"]
-          , msgRow "ReqTx" "received" (acReqTx c)
-          , msgRow "ReqSn" "sent (decided)" (acReqSnDecided c)
-          , msgRow "ReqSn" "received" (acReqSnRecv c)
-          , msgRow "ReqDec" "received" (acReqDec c)
-          , msgRow "(snapshots confirmed)" "" (acSnapConfirmed c)
-          ]
- where
-  peerAcks [] _ = div_ [styleInline_ "margin-left: 1rem;"] [italic "no peers yet"]
-  peerAcks peers byParty =
-    div_
-      [styleInline_ "margin: 0.1rem 0 0.2rem 1rem;"]
-      [ div_
-        [styleInline_ "display: flex; gap: 0.75rem; margin: 0.1rem 0; align-items: baseline;"]
-        [ span_
-            [styleInline_ "min-width: 14rem; font-family: ui-monospace, monospace; color: #335;"]
-            [text (ms (shortParty p))]
-        , span_
-            [styleInline_ "font-family: ui-monospace, monospace; font-weight: 600;"]
-            [text (ms (show (Map.findWithDefault 0 p byParty) :: Text))]
-        ]
-      | p <- peers
-      ]
-  msgRow name dir n =
-    div_
-      [styleInline_ "display: flex; gap: 0.75rem; margin: 0.15rem 0 0.15rem 1rem; align-items: baseline;"]
-      [ span_
-          [styleInline_ "min-width: 8rem; font-family: ui-monospace, monospace; color: #335;"]
-          [text name]
-      , span_ [styleInline_ "min-width: 8rem; color: #888;"] [text dir]
-      , span_
-          [styleInline_ "font-family: ui-monospace, monospace; font-weight: 600;"]
-          [text (ms (show n :: Text))]
-      ]
-
--- | Parties (peers) in the head, from the head parameters; empty before open.
-headParties :: HS.HeadState tx -> [Party]
-headParties = \case
-  HS.Open HS.OpenState{HS.parameters = ps} -> HP.parties ps
-  HS.Closed HS.ClosedState{HS.parameters = ps} -> HP.parties ps
-  HS.Idle _ -> []
-
--- | Render the currently-connected peer set as @host:port@, comma separated.
-connectedPeersText :: Set Host -> Text
-connectedPeersText hs
-  | Set.null hs = "none"
-  | otherwise = T.intercalate ", " (hostText <$> Set.toList hs)
-
-hostText :: Host -> Text
-hostText h = Net.hostname h <> ":" <> show (Net.port h)
-
--- | Running counts of chain/connectivity events and applied consensus
--- messages, accumulated by walking the history up to the cursor.
-data ActivityCounts = ActivityCounts
-  { acTicks :: Int
-  , acRollbacks :: Int
-  , acSynced :: Int
-  , acUnsynced :: Int
-  , acConnected :: Maybe Bool
-  -- ^ Latest network connectivity seen: 'Just True' connected, 'Just False'
-  -- disconnected, 'Nothing' if no connectivity event observed yet.
-  , acConnectedPeers :: Set Host
-  -- ^ Peers currently connected (PeerConnected adds, PeerDisconnected removes).
-  , acReqTx :: Int
-  , acReqSnRecv :: Int
-  , acReqSnDecided :: Int
-  , acAckByParty :: Map Party Int
-  -- ^ AckSn signatures recorded, grouped by the signing party.
-  , acReqDec :: Int
-  , acSnapConfirmed :: Int
-  }
-
-emptyActivity :: ActivityCounts
-emptyActivity =
-  ActivityCounts
-    { acTicks = 0
-    , acRollbacks = 0
-    , acSynced = 0
-    , acUnsynced = 0
-    , acConnected = Nothing
-    , acConnectedPeers = mempty
-    , acReqTx = 0
-    , acReqSnRecv = 0
-    , acReqSnDecided = 0
-    , acAckByParty = mempty
-    , acReqDec = 0
-    , acSnapConfirmed = 0
-    }
-
--- | The prefix of the *full* history (ticks included) up to and including the
--- cursor's event. Counts derived from this stay correct even while ticks are
--- hidden from the slider, which filters 'visibleHistory'.
-historyUpToCursor :: Model tx -> Vector (HistoryStep tx)
-historyUpToCursor m = case currentStep m of
-  Nothing -> V.empty
-  Just step ->
-    let eid = eventId (event step)
-     in V.takeWhile (\s -> eventId (event s) <= eid) (history m)
-
-countActivity :: Vector (HistoryStep tx) -> ActivityCounts
-countActivity = V.foldl' bump emptyActivity
- where
-  bump c step = case stateChanged (event step) of
-    TickObserved{} -> c{acTicks = acTicks c + 1}
-    ChainRolledBack{} -> c{acRollbacks = acRollbacks c + 1}
-    NodeSynced{} -> c{acSynced = acSynced c + 1}
-    NodeUnsynced{} -> c{acUnsynced = acUnsynced c + 1}
-    NetworkConnected{} -> c{acConnected = Just True}
-    NetworkDisconnected{} -> c{acConnected = Just False}
-    PeerConnected{peer} -> c{acConnectedPeers = Set.insert peer (acConnectedPeers c)}
-    PeerDisconnected{peer} -> c{acConnectedPeers = Set.delete peer (acConnectedPeers c)}
-    TransactionReceived{} -> c{acReqTx = acReqTx c + 1}
-    SnapshotRequested{} -> c{acReqSnRecv = acReqSnRecv c + 1}
-    SnapshotRequestDecided{} -> c{acReqSnDecided = acReqSnDecided c + 1}
-    PartySignedSnapshot{party} -> c{acAckByParty = Map.insertWith (+) party 1 (acAckByParty c)}
-    DecommitRecorded{} -> c{acReqDec = acReqDec c + 1}
-    SnapshotConfirmed{} -> c{acSnapConfirmed = acSnapConfirmed c + 1}
-    _ -> c
-
-viewState :: ToJSON (NodeState tx) => Model tx -> View (Action tx)
-viewState m = case currentStep m of
-  Nothing -> div_ [] []
-  Just step ->
-    div_
-      [styleInline_ panelStyle]
-      [ h2_ [] [text "node state after this event"]
-      , renderJson (stateAfter step)
-      ]
-
-panelStyle :: MisoString
-panelStyle =
-  "border: 1px solid #ddd; border-radius: 4px; padding: 0.75rem 1rem; \
-  \margin-bottom: 1rem; background: #fafafa;"
-
--- | The authoring panel. Hidden when 'ctx' is 'Nothing' (i.e. in Cardano
--- mode, where hand-crafting an 'Input Tx' as JSON is impractical).
-viewAuthoring :: ToJSON (Input tx) => Model tx -> View (Action tx)
-viewAuthoring m = case ctx m of
-  Nothing -> div_ [] []
-  Just _ ->
-    div_
-      [styleInline_ panelStyle]
-      [ h2_ [] [text "author input"]
-      , div_
-          [styleInline_ "color: #555; margin-bottom: 0.5rem;"]
-          [ text $
-              "Pick a preset or type a JSON-encoded Input tx, then click "
-                <> "Apply to call HeadLogic.update against the state at "
-                <> "the current cursor."
-          ]
-      , div_
-          [styleInline_ "display: flex; gap: 0.5rem; flex-wrap: wrap; margin-bottom: 0.5rem;"]
-          (presetButtons (presetInputs m))
-      , textarea_
-          [ rows_ "6"
-          , value_ (inputDraft m)
-          , onInput SetInputDraft
-          , styleInline_
-              "width: 100%; font-family: ui-monospace, monospace; \
-              \font-size: 0.85rem; box-sizing: border-box;"
-          ]
-          []
-      , div_
-          [styleInline_ "display: flex; gap: 0.5rem; align-items: center; margin-top: 0.5rem;"]
-          [ button_ [onClick ApplyDraft] [text "Apply"]
-          , case inputError m of
-              Nothing -> span_ [styleInline_ "color: #2a7;"] [text "ready"]
-              Just err -> span_ [styleInline_ "color: #a33;"] [text err]
-          ]
-      ]
- where
-  presetButtons :: [(MisoString, MisoString)] -> [View (Action tx)]
-  presetButtons = map $ \(lbl, json) ->
-    button_ [onClick (LoadPreset json)] [text lbl]
-
-presetInputs :: ToJSON (Input tx) => Model tx -> [(MisoString, MisoString)]
-presetInputs m = case ctx m of
-  Nothing -> []
-  Just c -> [(lbl, encodeAsMiso inp) | (lbl, inp) <- ctxPresets c]
- where
-  encodeAsMiso :: ToJSON a => a -> MisoString
-  encodeAsMiso = ms . TE.decodeUtf8 . BSL.toStrict . encodePretty
-
--- * Flow chart
-
--- | Macro states visible in the diagram, matching the spec's main-chain
--- diagram: Idle / Open / Closed / fanoutProgress / Final. fanoutProgress is
--- a synthetic node (no distinct 'HeadState' constructor backs it; we
--- classify a NodeState as 'SFanoutProgress' when we know a partial fanout is
--- in flight). Final is also synthetic: the implementation reverts the head
--- to Idle on 'HeadFannedOut', but the spec models the destination of
--- @fanout@ / @finalPartialFanout@ as a separate terminal node.
-data StateKind = SIdle | SOpen | SClosed | SFanoutProgress | SFinal
-  deriving stock (Eq, Show)
-
-stateKindOf :: NodeState tx -> StateKind
-stateKindOf n = case headState n of
-  HS.Idle{} -> SIdle
-  HS.Open{} -> SOpen
-  HS.Closed{} -> SClosed
-
--- | Counts of transitions and within-state events accumulated by walking the
--- history in order. We do not have access to the truly-initial 'NodeState'
--- here, so the very first event is treated as a self-loop on whatever state
--- it lands in (real Hydra logs start in 'Idle' and stay there for the first
--- few catch-up events, so this is rarely visible).
-data TransitionCounts = TransitionCounts
-  { idleToOpen :: Int
-  -- ^ \"init\" in spec terminology.
-  , openToClosed :: Int
-  -- ^ \"close\".
-  , idleSelf :: Int
-  , openSelf :: Int
-  , closedSelf :: Int
-  , openIncrement :: Int
-  -- ^ 'CommitFinalized' events: increment self-loop on Open.
-  , openDecrement :: Int
-  -- ^ 'DecommitFinalized' events: decrement self-loop on Open.
-  , closedContest :: Int
-  -- ^ 'HeadContested' events: contest self-loop on Closed.
-  , closedToFanoutProgress :: Int
-  -- ^ First 'HeadPartialFannedOut' in a fanout sequence: Closed -> fanoutProgress.
-  , fanoutProgressSelf :: Int
-  -- ^ Subsequent 'HeadPartialFannedOut' while already in a partial sequence:
-  -- fanoutProgress -> fanoutProgress.
-  , fanoutProgressToFinal :: Int
-  -- ^ 'HeadFannedOut' after one or more partials: fanoutProgress -> Final.
-  , closedToFinal :: Int
-  -- ^ 'HeadFannedOut' without any preceding partial: Closed -> Final
-  -- (single-tx fanout).
-  , unexpected :: Int
-  -- ^ Anything we did not expect (e.g. Closed -> Open). Surfacing this is
-  -- useful for spotting corrupt or surprising logs.
-  }
-  deriving stock (Eq, Show)
-
-emptyCounts :: TransitionCounts
-emptyCounts = TransitionCounts 0 0 0 0 0 0 0 0 0 0 0 0 0
-
--- | Walking state for 'countTransitions'. Tracks whether the head is
--- currently mid-partial-fanout so we can classify the subsequent
--- 'HeadFannedOut' as a fresh fanout vs. the closing step of a partial chain.
-data WalkState = WalkState
-  { wsPrevKind :: StateKind
-  , wsInPartial :: Bool
-  , wsCounts :: TransitionCounts
-  }
-
--- | Walk the history and return the full 'WalkState' (counts plus the
--- final 'wsInPartial' flag), so the view can read that flag to decide
--- whether the current macro state is 'SFanoutProgress' rather than 'SClosed'.
-walkHistory :: Vector (HistoryStep tx) -> WalkState
-walkHistory steps = case V.toList steps of
-  [] -> WalkState SIdle False emptyCounts
-  (s0 : rest) ->
-    let kind0 = stateKindOf (stateAfter s0)
-        ws0 = bump (WalkState kind0 False emptyCounts) kind0 s0
-     in foldl' stepFold ws0 rest
- where
-  stepFold :: WalkState -> HistoryStep tx -> WalkState
-  stepFold ws s = bump ws (stateKindOf (stateAfter s)) s
-
-  bump :: WalkState -> StateKind -> HistoryStep tx -> WalkState
-  bump ws curr step =
-    let prev = wsPrevKind ws
-        inPartial = wsInPartial ws
-        c = wsCounts ws
-     in case stateChanged (event step) of
-          CommitFinalized{} ->
-            ws{wsPrevKind = curr, wsCounts = c{openIncrement = openIncrement c + 1}}
-          DecommitFinalized{} ->
-            ws{wsPrevKind = curr, wsCounts = c{openDecrement = openDecrement c + 1}}
-          HeadContested{} ->
-            ws{wsPrevKind = curr, wsCounts = c{closedContest = closedContest c + 1}}
-          HeadPartialFannedOut{} ->
-            let c' =
-                  if inPartial
-                    then c{fanoutProgressSelf = fanoutProgressSelf c + 1}
-                    else c{closedToFanoutProgress = closedToFanoutProgress c + 1}
-             in ws{wsPrevKind = curr, wsInPartial = True, wsCounts = c'}
-          HeadFannedOut{} ->
-            let c' =
-                  if inPartial
-                    then c{fanoutProgressToFinal = fanoutProgressToFinal c + 1}
-                    else c{closedToFinal = closedToFinal c + 1}
-             in ws{wsPrevKind = curr, wsInPartial = False, wsCounts = c'}
-          _ ->
-            let c' = case (prev, curr) of
-                  (SIdle, SIdle) -> c{idleSelf = idleSelf c + 1}
-                  (SOpen, SOpen) -> c{openSelf = openSelf c + 1}
-                  (SClosed, SClosed) -> c{closedSelf = closedSelf c + 1}
-                  (SIdle, SOpen) -> c{idleToOpen = idleToOpen c + 1}
-                  (SOpen, SClosed) -> c{openToClosed = openToClosed c + 1}
-                  _ -> c{unexpected = unexpected c + 1}
-             in ws{wsPrevKind = curr, wsCounts = c'}
-
-viewFlow :: Model tx -> View (Action tx)
-viewFlow m =
-  let visible = V.take (cursor m + 1) (visibleHistory m)
-      ws = walkHistory visible
-      counts = wsCounts ws
-      -- Walking-state /before/ the cursor's own event: used to decide which
-      -- partial-fanout bucket the cursor's event lands in (and therefore
-      -- which arrow to highlight).
-      priorInPartial = wsInPartial (walkHistory (V.take (cursor m) (visibleHistory m)))
-      active = activeEdgeFor (currentStep m) priorInPartial
-      isE e = active == Just e
-      currentNode :: Maybe StateKind
-      currentNode = case currentStep m of
-        Nothing -> Nothing
-        Just step ->
-          Just $
-            if wsInPartial ws
-              then SFanoutProgress
-              else stateKindOf (stateAfter step)
-      is k = currentNode == Just k
-   in div_
-        [styleInline_ panelStyle]
-        [ h2_ [] [text "state transitions"]
-        , Svg.svg_
-            [ SvgA.viewBox_ "0 0 920 440"
-            , SvgA.width_ "100%"
-            , styleInline_ "max-width: 920px; display: block; margin: 0 auto;"
-            ]
-            ( markerDefs
-                : [ -- generic self-loops above the main row
-                    selfLoop 90 200 (idleSelf counts)
-                  , selfLoop 270 200 (openSelf counts)
-                  , selfLoopWithLabel 460 200 "contest" (closedContest counts) "#777" (isE EContest)
-                  , -- horizontal transitions across the main row
-                    arrow 135 200 225 200 (idleToOpen counts) "init" (isE EInit)
-                  , arrow 315 200 415 200 (openToClosed counts) "close" (isE EClose)
-                  , arrow 505 200 755 200 (closedToFinal counts) "fanout" (isE EFanout)
-                  , -- Open side-loops for increment / decrement
-                    sideLoopOnOpen 270 OnLeft "decrement" (openDecrement counts) "#c66" (isE EDecrement)
-                  , sideLoopOnOpen 270 OnRight "increment" (openIncrement counts) "#393" (isE EIncrement)
-                  , -- partial fanout: Closed -> fanoutProgress (diagonal down).
-                    -- Endpoints sit on the Closed circle (r=45) and the
-                    -- fanoutProgress ellipse (rx=75, ry=32) along the
-                    -- centre-to-centre line so the arrowhead touches both
-                    -- nodes cleanly.
-                    diagArrow 494 230 587 311 (closedToFanoutProgress counts) "partial fanout" "#a37" False (isE EClosedToFanoutProgress)
-                  , -- partial fanout: fanoutProgress self-loop (more partials)
-                    fanoutProgressSelfLoop 620 (fanoutProgressSelf counts) (isE EFanoutProgressSelf)
-                  , -- finalPartialFanout: fanoutProgress -> Final (diagonal up)
-                    diagArrow 656 312 765 228 (fanoutProgressToFinal counts) "final partial fanout" "#c83" True (isE EFinalPartialFanout)
-                  , -- nodes
-                    stateNode 90 200 "Idle" (is SIdle)
-                  , stateNode 270 200 "Open" (is SOpen)
-                  , stateNode 460 200 "Closed" (is SClosed)
-                  , fanoutProgressNode 620 340 (is SFanoutProgress)
-                  , stateNode 800 200 "Final" (is SFinal)
-                  ]
-                  <> [unexpectedNote (unexpected counts) | unexpected counts > 0]
-            )
-        ]
-
-data Side = OnLeft | OnRight
-
--- | Which animatable edge (if any) the cursor's event is currently taking.
-data ActiveEdge
-  = EInit
-  | EClose
-  | EContest
-  | EIncrement
-  | EDecrement
-  | EClosedToFanoutProgress
-  | EFanoutProgressSelf
-  | EFinalPartialFanout
-  | EFanout
-  deriving stock (Eq, Show)
-
--- | Decide which edge the event at the cursor is taking. The @priorInPartial@
--- argument is the walking-state's @wsInPartial@ flag BEFORE this event, used
--- to distinguish the first partial fanout (Closed -> fanoutProgress) from
--- subsequent partials (fanoutProgress self-loop), and similarly fresh fanout
--- (Closed -> Final) from the closing step of a partial sequence
--- (fanoutProgress -> Final).
-activeEdgeFor :: Maybe (HistoryStep tx) -> Bool -> Maybe ActiveEdge
-activeEdgeFor Nothing _ = Nothing
-activeEdgeFor (Just step) priorInPartial = case stateChanged (event step) of
-  HeadOpened{} -> Just EInit
-  HeadClosed{} -> Just EClose
-  HeadContested{} -> Just EContest
-  CommitFinalized{} -> Just EIncrement
-  DecommitFinalized{} -> Just EDecrement
-  HeadPartialFannedOut{}
-    | priorInPartial -> Just EFanoutProgressSelf
-    | otherwise -> Just EClosedToFanoutProgress
-  HeadFannedOut{}
-    | priorInPartial -> Just EFinalPartialFanout
-    | otherwise -> Just EFanout
-  _ -> Nothing
-
--- | Edge stroke colour: bright blue when this edge is the one the cursor's
--- event is currently traversing, otherwise the caller's chosen colour.
-activeStroke :: MisoString -> Bool -> MisoString
-activeStroke _ True = "#06c"
-activeStroke c False = c
-
--- | Edge stroke width: thicker when active.
-activeWidth :: MisoString -> Bool -> MisoString
-activeWidth _ True = "3.5"
-activeWidth w False = w
-
-markerDefs :: View (Action tx)
-markerDefs =
-  Svg.defs_
-    []
-    [ Svg.marker_
-        [ SvgA.id_ "arrowhead"
-        , SvgA.viewBox_ "0 0 10 10"
-        , SvgA.refX_ "9"
-        , SvgA.refY_ "5"
-        , SvgA.markerWidth_ "7"
-        , SvgA.markerHeight_ "7"
-        , SvgA.orient_ "auto-start-reverse"
-        ]
-        [ Svg.path_
-            [ SvgA.d_ "M0,0 L10,5 L0,10 z"
-            , SvgA.fill_ "#444"
-            ]
-            []
-        ]
-    ]
-
-stateNode :: Int -> Int -> MisoString -> Bool -> View (Action tx)
-stateNode cx cy nodeLabel isCurrent =
-  Svg.g_
-    []
-    [ Svg.circle_
-        [ SvgA.cx_ (mInt cx)
-        , SvgA.cy_ (mInt cy)
-        , SvgA.r_ "45"
-        , SvgA.fill_ (if isCurrent then "#cce5ff" else "#fff")
-        , SvgA.stroke_ "#444"
-        , SvgA.strokeWidth_ "2"
-        ]
-        []
-    , Svg.text_
-        [ SvgA.x_ (mInt cx)
-        , SvgA.y_ (mInt (cy + 5))
-        , SvgA.textAnchor_ "middle"
-        , SvgA.fontSize_ "18"
-        , SvgA.fontFamily_ "system-ui, sans-serif"
-        , SvgA.fill_ "#111"
-        ]
-        [text nodeLabel]
-    ]
-
-arrow :: Int -> Int -> Int -> Int -> Int -> MisoString -> Bool -> View (Action tx)
-arrow x1 y1 x2 y2 n lbl isActive =
-  Svg.g_
-    []
-    [ Svg.line_
-        [ SvgA.x1_ (mInt x1)
-        , SvgA.y1_ (mInt y1)
-        , SvgA.x2_ (mInt x2)
-        , SvgA.y2_ (mInt y2)
-        , SvgA.stroke_ (activeStroke "#444" isActive)
-        , SvgA.strokeWidth_ (activeWidth "2" isActive)
-        , SvgA.markerEnd_ "url(#arrowhead)"
-        ]
-        []
-    , Svg.text_
-        [ SvgA.x_ (mInt ((x1 + x2) `div` 2))
-        , SvgA.y_ (mInt (y1 - 18))
-        , SvgA.textAnchor_ "middle"
-        , SvgA.fontSize_ "11"
-        , SvgA.fontFamily_ "system-ui, sans-serif"
-        , SvgA.fill_ "#777"
-        ]
-        [text lbl]
-    , edgeLabel ((x1 + x2) `div` 2) (y1 - 4) n
-    ]
-
--- | A straight diagonal arrow with a label and count badge. Used for the
--- Closed -> fanoutProgress and fanoutProgress -> Final edges.
-diagArrow :: Int -> Int -> Int -> Int -> Int -> MisoString -> MisoString -> Bool -> Bool -> View (Action tx)
-diagArrow x1 y1 x2 y2 n lbl color dashed isActive =
-  let mx = (x1 + x2) `div` 2
-      my = (y1 + y2) `div` 2
-      lineAttrs =
-        [ SvgA.x1_ (mInt x1)
-        , SvgA.y1_ (mInt y1)
-        , SvgA.x2_ (mInt x2)
-        , SvgA.y2_ (mInt y2)
-        , SvgA.stroke_ (activeStroke color isActive)
-        , SvgA.strokeWidth_ (activeWidth "1.8" isActive)
-        , SvgA.markerEnd_ "url(#arrowhead)"
-        ]
-          <> [SvgA.strokeDasharray_ "5 4" | dashed && not isActive]
-   in Svg.g_
-        []
-        [ Svg.line_ lineAttrs []
-        , Svg.text_
-            [ SvgA.x_ (mInt mx)
-            , SvgA.y_ (mInt (my - 6))
-            , SvgA.textAnchor_ "middle"
-            , SvgA.fontSize_ "11"
-            , SvgA.fontFamily_ "system-ui, sans-serif"
-            , SvgA.fill_ color
-            ]
-            [text lbl]
-        , edgeLabel mx (my + 14) n
-        ]
-
--- | Self-loop drawn below the fanoutProgress ellipse for the @partialFanout@
--- transitions that keep the head in fanoutProgress.
-fanoutProgressSelfLoop :: Int -> Int -> Bool -> View (Action tx)
-fanoutProgressSelfLoop cx n isActive =
-  Svg.g_
-    []
-    [ Svg.path_
-        [ SvgA.d_
-            ( "M "
-                <> mInt (cx - 25)
-                <> " 372 C "
-                <> mInt (cx - 70)
-                <> " 425 "
-                <> mInt (cx + 70)
-                <> " 425 "
-                <> mInt (cx + 25)
-                <> " 372"
-            )
-        , SvgA.fill_ "none"
-        , SvgA.stroke_ (activeStroke "#a37" isActive)
-        , SvgA.strokeWidth_ (activeWidth "1.8" isActive)
-        , SvgA.markerEnd_ "url(#arrowhead)"
-        ]
-        []
-    , Svg.text_
-        [ SvgA.x_ (mInt cx)
-        , SvgA.y_ "418"
-        , SvgA.textAnchor_ "middle"
-        , SvgA.fontSize_ "12"
-        , SvgA.fontFamily_ "system-ui, sans-serif"
-        , SvgA.fill_ "#a37"
-        ]
-        [text "partial fanout"]
-    , edgeLabel cx 398 n
-    ]
-
--- | An elliptical node for fanoutProgress (drawn wider than the circular
--- states so the longer label fits inside).
-fanoutProgressNode :: Int -> Int -> Bool -> View (Action tx)
-fanoutProgressNode cx cy isCurrent =
-  Svg.g_
-    []
-    [ Svg.ellipse_
-        [ SvgA.cx_ (mInt cx)
-        , SvgA.cy_ (mInt cy)
-        , SvgA.rx_ "75"
-        , SvgA.ry_ "32"
-        , SvgA.fill_ (if isCurrent then "#cce5ff" else "#fff")
-        , SvgA.stroke_ "#444"
-        , SvgA.strokeWidth_ "2"
-        ]
-        []
-    , Svg.text_
-        [ SvgA.x_ (mInt cx)
-        , SvgA.y_ (mInt (cy + 5))
-        , SvgA.textAnchor_ "middle"
-        , SvgA.fontSize_ "15"
-        , SvgA.fontFamily_ "system-ui, sans-serif"
-        , SvgA.fill_ "#111"
-        ]
-        [text "fanoutProgress"]
-    ]
-
--- | A curved arc from the bottom of the Open node out to one side and back,
--- drawn below the main horizontal row so it never crosses the open/close
--- arrows. Used for the increment ('CommitFinalized') and decommit
--- ('DecommitFinalized') self-transitions, which keep the head in Open but
--- represent real protocol events worth surfacing distinctly.
--- | A self-loop drawn below the Open node, jutting out to one side. Used
--- for the increment and decrement edges. Anchors at ~45deg below the node
--- centre so the arc starts/ends on the node boundary, not floating off it.
-sideLoopOnOpen :: Int -> Side -> MisoString -> Int -> MisoString -> Bool -> View (Action tx)
-sideLoopOnOpen cx side lbl n color isActive =
-  let sign :: Int
-      sign = case side of OnRight -> 1; OnLeft -> -1
-      -- Open node is at (cx, 200) with radius 45. Anchor the arc start and
-      -- end on the bottom arc of the node (~30deg apart) so both ends touch
-      -- the boundary cleanly. Curve out to ~70px sideways and ~50px down.
-      x0 = cx + sign * 22
-      y0 = 240
-      x1 = cx + sign * 70
-      y1 = 270
-      x2 = cx + sign * 70
-      y2 = 290
-      x3 = cx + sign * 40
-      y3 = 222
-      d =
-        "M "
-          <> mInt x0
-          <> " "
-          <> mInt y0
-          <> " C "
-          <> mInt x1
-          <> " "
-          <> mInt y1
-          <> " "
-          <> mInt x2
-          <> " "
-          <> mInt y2
-          <> " "
-          <> mInt x3
-          <> " "
-          <> mInt y3
-      labelX = cx + sign * 70
-      labelY = 280
-   in Svg.g_
-        []
-        [ Svg.path_
-            [ SvgA.d_ d
-            , SvgA.fill_ "none"
-            , SvgA.stroke_ (activeStroke color isActive)
-            , SvgA.strokeWidth_ (activeWidth "1.8" isActive)
-            , SvgA.markerEnd_ "url(#arrowhead)"
-            ]
-            []
-        , Svg.text_
-            [ SvgA.x_ (mInt labelX)
-            , SvgA.y_ (mInt (labelY - 22))
-            , SvgA.textAnchor_ "middle"
-            , SvgA.fontSize_ "12"
-            , SvgA.fontFamily_ "system-ui, sans-serif"
-            , SvgA.fill_ color
-            ]
-            [text lbl]
-        , edgeLabel labelX (labelY - 8) n
-        ]
-
-selfLoop :: Int -> Int -> Int -> View (Action tx)
-selfLoop cx cy n =
-  Svg.g_
-    []
-    [ Svg.path_
-        [ -- Anchors are at ~45 degrees on the top arc of a radius-45 circle
-          -- (so the start/end lie on the node edge, not floating outside).
-          SvgA.d_
-            ( "M "
-                <> mInt (cx - 32)
-                <> " "
-                <> mInt (cy - 32)
-                <> " C "
-                <> mInt (cx - 60)
-                <> " "
-                <> mInt (cy - 100)
-                <> " "
-                <> mInt (cx + 60)
-                <> " "
-                <> mInt (cy - 100)
-                <> " "
-                <> mInt (cx + 32)
-                <> " "
-                <> mInt (cy - 32)
-            )
-        , SvgA.fill_ "none"
-        , SvgA.stroke_ "#888"
-        , SvgA.strokeWidth_ "1.5"
-        , SvgA.strokeDasharray_ "4 3"
-        , SvgA.markerEnd_ "url(#arrowhead)"
-        ]
-        []
-    , edgeLabel cx (cy - 80) n
-    ]
-
--- | Like 'selfLoop' but with a coloured text label sitting next to the
--- count, used for the 'contest' self-loop on Closed.
-selfLoopWithLabel :: Int -> Int -> MisoString -> Int -> MisoString -> Bool -> View (Action tx)
-selfLoopWithLabel cx cy lbl n color isActive =
-  Svg.g_
-    []
-    [ Svg.path_
-        [ SvgA.d_
-            ( "M "
-                <> mInt (cx - 32)
-                <> " "
-                <> mInt (cy - 32)
-                <> " C "
-                <> mInt (cx - 60)
-                <> " "
-                <> mInt (cy - 100)
-                <> " "
-                <> mInt (cx + 60)
-                <> " "
-                <> mInt (cy - 100)
-                <> " "
-                <> mInt (cx + 32)
-                <> " "
-                <> mInt (cy - 32)
-            )
-        , SvgA.fill_ "none"
-        , SvgA.stroke_ (activeStroke color isActive)
-        , SvgA.strokeWidth_ (activeWidth "1.5" isActive)
-        , SvgA.markerEnd_ "url(#arrowhead)"
-        ]
-        []
-    , Svg.text_
-        [ SvgA.x_ (mInt cx)
-        , SvgA.y_ (mInt (cy - 100))
-        , SvgA.textAnchor_ "middle"
-        , SvgA.fontSize_ "12"
-        , SvgA.fontFamily_ "system-ui, sans-serif"
-        , SvgA.fill_ color
-        ]
-        [text lbl]
-    , edgeLabel cx (cy - 80) n
-    ]
-
-edgeLabel :: Int -> Int -> Int -> View (Action tx)
-edgeLabel cx cy n =
-  Svg.g_
-    []
-    [ Svg.rect_
-        [ SvgA.x_ (mInt (cx - 18))
-        , SvgA.y_ (mInt (cy - 13))
-        , SvgA.width_ "36"
-        , SvgA.height_ "20"
-        , SvgA.rx_ "10"
-        , SvgA.fill_ "#fff"
-        , SvgA.stroke_ "#ccc"
-        ]
-        []
-    , Svg.text_
-        [ SvgA.x_ (mInt cx)
-        , SvgA.y_ (mInt (cy + 3))
-        , SvgA.textAnchor_ "middle"
-        , SvgA.fontSize_ "13"
-        , SvgA.fontFamily_ "system-ui, sans-serif"
-        , SvgA.fill_ "#333"
-        ]
-        [text (ms (show n :: Text))]
-    ]
-
-unexpectedNote :: Int -> View (Action tx)
-unexpectedNote n =
-  Svg.text_
-    [ SvgA.x_ "360"
-    , SvgA.y_ "210"
-    , SvgA.textAnchor_ "middle"
-    , SvgA.fontSize_ "12"
-    , SvgA.fill_ "#a33"
-    ]
-    [text $ "(+" <> ms (show n :: Text) <> " unexpected transitions)"]
-
-mInt :: Int -> MisoString
-mInt n = ms (show n :: Text)
-
-currentStep :: Model tx -> Maybe (HistoryStep tx)
-currentStep m = visibleHistory m V.!? cursor m
-
-renderJson :: ToJSON a => a -> View (Action tx)
-renderJson x =
-  pre_
-    [ styleInline_
-        "background: white; border: 1px solid #eee; border-radius: 3px; \
-        \padding: 0.5rem; overflow-x: auto; font-size: 0.85rem; \
-        \max-height: 30rem; overflow-y: auto;"
-    ]
-    [ code_ [] [text (ms (jsonText x))]
-    ]
-
-jsonText :: ToJSON a => a -> Text
-jsonText = TE.decodeUtf8 . BSL.toStrict . encodePretty
